# Avoid repeatedly encoding method name strings in protocol