--- a/src/SignalR/server/Core/src/Internal/Utf8HashLookup.cs
+++ b/src/SignalR/server/Core/src/Internal/Utf8HashLookup.cs
@@ -13,43 +13,93 @@ namespace Microsoft.AspNetCore.SignalR.Internal;
 internal sealed class Utf8HashLookup
 {
     private int[] _buckets;
+    private int[] _caseSensitiveBuckets;
     private Slot[] _slots;
     private int _count;
+
+    private const int HashCodeMask = 0x7fffffff;
 
     internal Utf8HashLookup()
     {
         _buckets = new int[7];
+        _caseSensitiveBuckets = new int[7];
         _slots = new Slot[7];
     }
 
     internal void Add(string value)
     {
-        var key = Encoding.UTF8.GetBytes(value);
-        var hashCode = GetKeyHashCode(key);
-
         if (_count == _slots.Length)
         {
             Resize();
         }
 
-        int index = _count;
+        int slotIndex = _count;
         _count++;
 
-        int bucket = hashCode % _buckets.Length;
-        _slots[index].hashCode = hashCode;
-        _slots[index].key = key;
-        _slots[index].value = value;
-        _slots[index].next = _buckets[bucket] - 1;
-        _buckets[bucket] = index + 1;
+        var encodedValue = Encoding.UTF8.GetBytes(value);
+        var hashCode = GetHashCode(value.AsSpan());
+        var caseSensitiveHashCode = GetCaseSensitiveHashCode(encodedValue);
+        int bucketIndex = hashCode % _buckets.Length;
+        int caseSensitiveBucketIndex = caseSensitiveHashCode % _caseSensitiveBuckets.Length;
+
+        _slots[slotIndex].hashCode = hashCode;
+        _slots[slotIndex].caseSensitiveHashCode = caseSensitiveHashCode;
+
+        _slots[slotIndex].value = value;
+        _slots[slotIndex].encodedValue = encodedValue;
+
+
+        _slots[slotIndex].next = _buckets[bucketIndex] - 1;
+        _slots[slotIndex].caseSensitiveNext = _caseSensitiveBuckets[caseSensitiveBucketIndex] - 1;
+
+        _buckets[bucketIndex] = slotIndex + 1;
+        _caseSensitiveBuckets[caseSensitiveBucketIndex] = slotIndex + 1;
     }
 
-    internal bool TryGetValue(ReadOnlySpan<byte> key, [MaybeNullWhen(false), AllowNull] out string value)
+    internal bool TryGetValue(ReadOnlySpan<byte> encodedValue, [MaybeNullWhen(false), AllowNull] out string value)
     {
-        var hashCode = GetKeyHashCode(key);
+        var caseSensitiveHashCode = GetCaseSensitiveHashCode(encodedValue);
+
+        for (var i = _caseSensitiveBuckets[caseSensitiveHashCode % _caseSensitiveBuckets.Length] - 1; i >= 0; i = _slots[i].caseSensitiveNext)
+        {
+            if (_slots[i].caseSensitiveHashCode == caseSensitiveHashCode && encodedValue.SequenceEqual(_slots[i].encodedValue.Span))
+            {
+                value = _slots[i].value;
+                return true;
+            }
+        }
+
+        // If we cannot find a case-sensitive match, we transcode the encodedValue to a stackalloced UTF16 string
+        // and do an OrdinalIgnoreCase comparison.
+        return TryGetValueSlow(encodedValue, out value);
+    }
+
+    private bool TryGetValueSlow(ReadOnlySpan<byte> encodedValue, [MaybeNullWhen(false), AllowNull] out string value)
+    {
+        const int StackAllocThreshold = 128;
+
+        char[]? pooled = null;
+        var count = Encoding.UTF8.GetCharCount(encodedValue);
+        var chars = count <= StackAllocThreshold ?
+            stackalloc char[StackAllocThreshold] :
+            (pooled = ArrayPool<char>.Shared.Rent(count));
+        var encoded = Encoding.UTF8.GetChars(encodedValue, chars);
+        var hasValue = TryGetValueFromChars(chars[..encoded], out value);
+        if (pooled is not null)
+        {
+            ArrayPool<char>.Shared.Return(pooled);
+        }
+
+        return hasValue;
+    }
+
+    private bool TryGetValueFromChars(ReadOnlySpan<char> key, [MaybeNullWhen(false), AllowNull] out string value)
+    {
+        var hashCode = GetHashCode(key);
 
         for (var i = _buckets[hashCode % _buckets.Length] - 1; i >= 0; i = _slots[i].next)
         {
-            if (_slots[i].hashCode == hashCode && key.SequenceEqual(_slots[i].key.Span))
+            if (_slots[i].hashCode == hashCode && key.Equals(_slots[i].value, StringComparison.OrdinalIgnoreCase))
             {
                 value = _slots[i].value;
                 return true;
@@ -60,34 +110,52 @@ internal sealed class Utf8HashLookup
         return false;
     }
 
-    private static int GetKeyHashCode(ReadOnlySpan<byte> key)
+    private static int GetHashCode(ReadOnlySpan<char> value) =>
+        HashCodeMask & string.GetHashCode(value, StringComparison.OrdinalIgnoreCase);
+
+    private static int GetCaseSensitiveHashCode(ReadOnlySpan<byte> encodedValue)
     {
         var hashCode = new HashCode();
-        hashCode.AddBytes(key);
-        return hashCode.ToHashCode();
+        hashCode.AddBytes(encodedValue);
+        return HashCodeMask & hashCode.ToHashCode();
     }
 
     private void Resize()
     {
         var newSize = checked(_count * 2 + 1);
-        var newBuckets = new int[newSize];
         var newSlots = new Slot[newSize];
+
+        var newBuckets = new int[newSize];
+        var newCaseSensitiveBuckets = new int[newSize];
+
         Array.Copy(_slots, newSlots, _count);
+
         for (int i = 0; i < _count; i++)
         {
             int bucket = newSlots[i].hashCode % newSize;
             newSlots[i].next = newBuckets[bucket] - 1;
             newBuckets[bucket] = i + 1;
+
+            int caseSensitiveBucket = newSlots[i].caseSensitiveHashCode % newSize;
+            newSlots[i].caseSensitiveNext = newCaseSensitiveBuckets[caseSensitiveBucket] - 1;
+            newCaseSensitiveBuckets[caseSensitiveBucket] = i + 1;
         }
-        _buckets = newBuckets;
+
         _slots = newSlots;
+
+        _buckets = newBuckets;
+        _caseSensitiveBuckets = newCaseSensitiveBuckets;
     }
 
     private struct Slot
     {
         internal int hashCode;
-        internal int next;
-        internal Memory<byte> key;
+        internal int caseSensitiveHashCode;
+
         internal string value;
+        internal Memory<byte> encodedValue;
+
+        internal int next;
+        internal int caseSensitiveNext;
     }
 }