This is a follow up to https://github.com/dotnet/aspnetcore/pull/41465 and specifically https://github.com/dotnet/aspnetcore/pull/41465#discussion_r869991091.

I reran the `StringLookup` and `Uf8Lookup` (sic) scenarios from the original PR's `StringLookupBenchmark` after updating the `Payload` to match the expected casing. This yields a perf improvement over repeatedly allocating the string and looking it up in a dictionary instead of a perf regression like before:

### Before

|       Method |      Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------- |----------:|---------:|---------:|-------:|----------:|
| StringLookup |  92.02 ns | 1.824 ns | 1.523 ns | 0.0030 |      32 B |
|    Uf8Lookup | 104.24 ns | 0.286 ns | 0.239 ns |      - |         - |

### After

|       Method |     Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------- |---------:|---------:|---------:|-------:|----------:|
| StringLookup | 90.15 ns | 0.505 ns | 0.473 ns | 0.0030 |      32 B |
|   Uf8Lookup | 77.74 ns | 0.457 ns | 0.428 ns |      - |         - |

It also avoids any allocations in the common case where the casing matches. I expect mismatched casing between the server and client is really uncommon now that we don't generate automatic client proxies with different casing. The targets are just strings on the client and the server, so it doesn't make sense to mix up the casing. And even if the casing doesn't match, everything still works as before.