@davidfowl 2022-05-11
As discussed offline, this should support case insensitive lookup even if it's the fallback situation.