@halter73 2022-05-12
I updated this to fall back to the utf16-based case-insensitive comparison when there's not an exact ordinal match.

On my machine, the performance is still just under 80 ns per operation when there's an ordinal match vs 105 ns before this change. When there's a case-insensitive match perf goes from about 105 ns before to 110 ns after because of the extra SequenceEqual comparison, but now there's no allocations in either case.

the same when there's an ordinal match but about