@davidfowl 2022-05-12
PS your benchmarks don't have a case insensitive comparison.